+++ CHANGELOG.rst
@@ -38,6 +38,8 @@ in development
 * Ability to partition sensors across sensor nodes using various partition schemes. (new-feature)
 * Add ability to use action context params as action params in meta. (new-feature)
 * Fix a bug in action container where rendering params was done twice. (bug-fix)
+* Move /exp/actionalias/ and /exp/aliasexecution to /v1/actionalias/ and /v1/aliasexecution/
+  respectively. (upgrade)
 
 0.11.4 - June 30, 2015
 ----------------------